requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1
lxml==4.9.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
httpx==0.25.1
playwright==1.40.0
aiohttp==3.9.1
pandas==2.1.3
apscheduler==3.10.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
pytesseract==0.3.10
Pillow==10.1.0
